Add entry field history and improve field displays
Multiple updates were made to improve the functionality and display of entry fields. There is now a history available for the entry fields which implements both EntryFieldHistoryItem and EntryGroupHistoryItem. Adjustments were also made to better handle the layout of password details within EntryField and reduce comment-out code that was polluting the files.