Add copy-to-clipboard functionality and update application version
This commit introduces a "CopyToClipboard" component that allows users to copy different kinds of information displayed in the application, such as notes, passwords, and links. This will improve the user experience by simplifying the process of copying and pasting these details. The application version in 'Package.appxmanifest' has also been updated.